@@ -3,34 +3,35 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@hookform/resolvers": "^1.1.1",
-    "@testing-library/jest-dom": "^5.11.5",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "@hookform/resolvers": "^2.8.8",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
     "awesome-react-icons": "^1.4.8",
-    "axios": "^0.21.0",
-    "bootstrap": "^4.5.3",
-    "chart.js": "^2.9.4",
+    "axios": "^0.26.1",
+    "bootstrap": "^5.1.3",
+    "chart.js": "^3.7.1",
     "datapicker": "^1.0.0",
     "hook-form": "0.0.1",
-    "mdbreact": "^5.0.1",
-    "moment": "^2.29.1",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.4.0",
-    "react-calendar": "^3.1.0",
-    "react-dom": "^17.0.1",
+    "mdbreact": "^5.2.0",
+    "moment": "^2.29.3",
+    "react": "^18.0.0",
+    "react-bootstrap": "^2.3.0",
+    "react-calendar": "^3.7.0",
+    "react-dom": "^18.0.0",
     "react-form": "^4.0.1",
     "react-google-recaptcha": "^2.1.0",
-    "react-hook-form": "^6.12.2",
-    "react-minimal-side-navigation": "^1.8.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.0",
-    "react-yup": "^1.19.2",
+    "react-hook-form": "^7.30.0",
+    "react-icons": "^4.3.1",
+    "react-minimal-side-navigation": "^1.9.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "react-yup": "^1.23.0",
     "sha1": "^1.1.1",
-    "styled-components": "^5.2.1",
+    "styled-components": "^5.3.5",
     "sweetalert": "^2.1.2",
-    "web-vitals": "^0.2.4",
-    "yup": "^0.32.6"
+    "web-vitals": "^2.1.4",
+    "yup": "^0.32.11"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -55,5 +56,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "sass": "^1.50.1"
   }
 }
