@@ -1,17 +1,5 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
+import  {createRoot} from 'react-dom/client';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
-
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+const container = document.getElementById('app');
+const root = createRoot(container);
+root.render(<App tab="home" />);
